some fixes in market client and coinsigning

+++ crypto_ops/crypto/sha_src/sha512.cpp
@@ -120,7 +120,7 @@ std::string sha512 (const std::string &input) {
 
 	char buf[2 * SHA512::DIGEST_SIZE + 1];
 	buf[2 * SHA512::DIGEST_SIZE] = 0;
-	for (int i = 0; i < SHA512::DIGEST_SIZE; ++i) {
+	for (size_t i = 0; i < SHA512::DIGEST_SIZE; ++i) {
 		sprintf(buf + i * 2, "%02x", digest[i]);
 	}
 	return std::string(buf);
